Use IO::Socket::INET6
IO::Socket(::INET) doesn't work in IPv6 environments,
leading to failures of connectivity when a provided
hostname resolves to an AAAA record, as there is no
fallback to the IPv4 A record. IO::Socket::INET6 fixes
all that by providing both IPv4 and IPv6.

One could also use IO::Socket::IP, but it is not as
commonly packaged by distros.